change to int

+++ homework_6/include/Image.hpp
@@ -13,8 +13,8 @@ class Image {
   [[nodiscard]] int rows() const;
   [[nodiscard]] int cols() const;
 
-  [[nodiscard]] uint8_t at(int row, int col) const;
-  [[nodiscard]] uint8_t& at(int row, int col);
+  [[nodiscard]] int at(int row, int col) const;
+  [[nodiscard]] int& at(int row, int col);
 
   bool FillFromPgm(const std::string& file_name);
   void WriteToPgm(const std::string& file_name);
@@ -27,16 +27,16 @@ class Image {
  private:
   int rows_ = 0;
   int cols_ = 0;
-  uint8_t max_val_ = 0;
-  std::vector<uint8_t> data_;
+  int max_val_ = 0;
+  std::vector<int> data_;
   [[nodiscard]] static float mapping(size_t numb, size_t max);
   [[nodiscard]] std::vector<size_t> createCompleteHistogram() const;
   static std::vector<int> createReducedHistogram(
       int bins, size_t bin_size, const std::vector<size_t>& full_histogram);
   [[nodiscard]] static std::vector<float> createNormalizedHistogram(
       size_t fsize, const std::vector<int>& reduced_histogram);
-  void createBigPixel(size_t scale, std::vector<uint8_t>& new_data, size_t col,
-                      size_t row, uint8_t pixel_value) const;
+  void createBigPixel(size_t scale, std::vector<int>& new_data, size_t col,
+                      size_t row, int pixel_value) const;
 };
 
 }  // namespace igg
